+++ models/__init__.py
@@ -1,3 +1,4 @@
 from . import product_template_attribute_value
 from . import product_product
+from . import product_template
 from . import sale_order_line